+++ CHANGELOG.md
@@ -23,6 +23,8 @@ and this project adheres to
   [#3887](https://github.com/OpenFn/lightning/issues/3887)
 - Mix task to merge project state files without database access
   [#3615](https://github.com/OpenFn/lightning/issues/3615)
+- Support Undo/Redo commands in the collab editor
+  [#3712](https://github.com/OpenFn/lightning/issues/3712)
 
 ### Changed